// src/domains/pokedex/nature/enums/nature.enum.ts
export enum NatureEnum {
    HARDY = "hardy",
    LONELY = "lonely",
    BRAVE = "brave",
    ADAMANT = "adamant",
    NAUGHTY = "naughty",
    BOLD = "bold",
    DOCILE = "docile",
    RELAXED = "relaxed",
    IMPISH = "impish",
    LAX = "lax",
    TIMID = "timid",
    HASTY = "hasty",
    SERIOUS = "serious",
    JOLLY = "jolly",
    NAIVE = "naive",
    MODEST = "modest",
    MILD = "mild",
    QUIET = "quiet",
    BASHFUL = "bashful",
    RASH = "rash",
    CALM = "calm",
    GENTLE = "gentle",
    SASSY = "sassy",
    CAREFUL = "careful",
    QUIRKY = "quirky",
}